requirements.txt
asgiref==3.4.1
cached-property==1.5.2
click==8.0.3
Django==3.2
django-admin-display==1.3.0
graphql-core==3.1.7
packaging==21.3
Pygments==2.11.2
PyJWT==2.3.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2021.3
sentinel==0.3.0
six==1.16.0
sqlparse==0.4.2
strawberry-django-jwt==0.1.10
strawberry-graphql==0.93.22
strawberry-graphql-django==0.2.5
typing_extensions==4.0.1
